+++ NEWS.md
@@ -5,6 +5,7 @@
 * `typical_value()` gets a "zero"-option for the `fun`-argument.
 * Changes to `icc()`, which used `stats::sigma()` and thus required R-version 3.3 or higher. Now should depend on R 3.2 again.
 * `se()` now also supports `stanreg` and `stanfit` objects.
+* `hdi()` now also supports `stanfit`-objects.
 
 ## New functions